Add DocOps container images for Python
The setup for GitHub devcontainers and GitHub Actions workflows both
necessitate building Docker images. These Docker images need to be kept
in sync to ensure that the development and CI/CD environments are
identical

Repeating the same Docker configuration violates the DRY principle, so
this PR is my first attempt to build a DoiT DocOps specific set of
Docker images for use with Python applications.

The Docker images are built and published to this project's [packages
repository][packages] every time a new commit is made to the `main`
branch. For now, I have decided against the idea of using tags for this
repository, because I imagine it will eventually be home to a number of
different and unrelated images.

[packages]: https://github.com/orgs/doitintl/packages?repo_name=docops-containers