Improve devcontainer setup
Instead of depending on the old [docops-containers][1] images, this
change modifies the setup so that the project-local devcontainer
definition sets up everything it needs. Specifically, the image now
includes a full Homebrew installation.

Because it takes a while to install Homebrew, I have added a GitHub
Actions workflow to pre-build the devcontainer every time a change is
made to the devcontainer definition on the main branch.

To bootstrap this setup, the `devcontainer.json` file needs to be
configured to build its own image.

However, once a pre-built image has been published, the
`devcontainer.json` file can be modified to pull it directly from
GitHub.

[1]: https://github.com/doitintl/docops-containers