First rework of the DevOps containers
Instead of building containers with different Python versions, I am
going to build two variants: one for developers (`dev`) to use with VS
Code, and one for the GitHub-hosted runner (`runner`), with a
slimmed-down image (e.g., remove the Homebrew machinery, which takes up
a lot of space).

The `dev` variant is the base variant, and this commit modifies the
project to build that variant. The `runner` variant will be created in a
subsequent commit and will use the `dev` variant as its base.

This commit also:

- Simplifies the `.devcontainer` setup. I mistakenly thought I needed to
  use the "Docker in Docker" image to build containers. In fact, you can
  build containers with GitHub Actions without any special setup.

- Updates other miscellaneous files to be more in line with changes in
  other DocOps repositories.

- Added a section to the `LICENCE` file. A significant portion of these
  changes are derived from work I have done in my personal time on
  various Bottle Garden projects.