Add experimental Zapier hook for RTD
The new `slack/rtd/zapier.py` file is a first attempt at handling
incoming webhooks from RTD. When a webhook is received, this script
generates a payload that can be sent to a Slack application, resulting
in a channel notification.

Additionally, this commit adds some minor quality of life improvements
for this repository.